replace usage of USpan<T> with Span<T>

+++ core/Extensions/SerializableExtensions.cs
@@ -11,7 +11,7 @@
         /// </summary>
         public unsafe static T Clone<T>(this T self) where T : unmanaged, ISerializable
         {
-            using ByteWriter writer = new((uint)sizeof(T));
+            using ByteWriter writer = new(sizeof(T));
             self.Write(writer);
             using ByteReader reader = new(writer);
             return reader.ReadObject<T>();
